use MultiStageRequestIdGenerator in v1 engine
